GetAccount refactoring

+++ src/libData/AccountData/AccountStore.cpp
@@ -37,9 +37,10 @@ using namespace Contract;
 AccountStore::AccountStore() {
   m_accountStoreTemp = make_unique<AccountStoreTemp>(*this);
 
-  if (ENABLE_SC) {
+  if (ENABLE_SC && !LOOKUP_NODE_MODE) {
     /// Scilla IPC Server
     /// clear path
+    lock_guard<mutex> g(m_mutexIPC);
     boost::filesystem::remove_all(SCILLA_IPC_SOCKET_PATH);
     m_scillaIPCServerConnector =
         make_unique<jsonrpc::UnixDomainSocketServer>(SCILLA_IPC_SOCKET_PATH);
@@ -61,6 +62,7 @@ AccountStore::AccountStore() {
 
 AccountStore::~AccountStore() {
   // boost::filesystem::remove_all("./state");
+  lock_guard<mutex> g(m_mutexIPC);
   if (m_scillaIPCServer != nullptr) {
     m_scillaIPCServer->StopListening();
   }
@@ -71,37 +73,31 @@ void AccountStore::Init() {
 
   InitSoft();
 
-  lock_guard<mutex> g(m_mutexDB);
-
   ContractStorage2::GetContractStorage().Reset();
-  m_db.ResetDB();
+  AccountStoreTrie<OverlayDB,
+                   unordered_map<Address, std::shared_ptr<Account>>>::ResetDB();
 }
 
 void AccountStore::InitSoft() {
   LOG_MARKER();
 
-  unique_lock<shared_timed_mutex> g(m_mutexPrimary);
-
-  AccountStoreTrie<OverlayDB, unordered_map<Address, Account>>::Init();
+  AccountStoreTrie<OverlayDB,
+                   unordered_map<Address, std::shared_ptr<Account>>>::Init();
 
   InitRevertibles();
 
   InitTemp();
 }
 
-bool AccountStore::RefreshDB() {
-  LOG_MARKER();
-  lock_guard<mutex> g(m_mutexDB);
-  return m_db.RefreshDB();
-}
-
 void AccountStore::InitTemp() {
   LOG_MARKER();
 
-  lock_guard<mutex> g(m_mutexDelta);
-
   m_accountStoreTemp->Init();
-  m_stateDeltaSerialized.clear();
+
+  {
+    unique_lock<shared_timed_mutex> g(m_mutexDelta);
+    m_stateDeltaSerialized.clear();
+  }
 
   ContractStorage2::GetContractStorage().InitTempState();
 }
@@ -109,10 +105,12 @@ void AccountStore::InitTemp() {
 void AccountStore::InitRevertibles() {
   LOG_MARKER();
 
-  lock_guard<mutex> g(m_mutexRevertibles);
+  {
+    lock_guard<mutex> g(m_mutexRevertibles);
 
-  m_addressToAccountRevChanged.clear();
-  m_addressToAccountRevCreated.clear();
+    m_addressToAccountRevChanged.clear();
+    m_addressToAccountRevCreated.clear();
+  }
 
   ContractStorage2::GetContractStorage().InitRevertibles();
 }
@@ -124,18 +122,16 @@ AccountStore& AccountStore::GetInstance() {
 
 bool AccountStore::Serialize(bytes& src, unsigned int offset) const {
   LOG_MARKER();
-  shared_lock<shared_timed_mutex> lock(m_mutexPrimary);
   return AccountStoreTrie<
-      dev::OverlayDB, std::unordered_map<Address, Account>>::Serialize(src,
-                                                                       offset);
+      dev::OverlayDB,
+      std::unordered_map<Address, std::shared_ptr<Account>>>::Serialize(src,
+                                                                        offset);
 }
 
 bool AccountStore::Deserialize(const bytes& src, unsigned int offset) {
   LOG_MARKER();
 
   this->Init();
-
-  unique_lock<shared_timed_mutex> g(m_mutexPrimary);
 
   if (!Messenger::GetAccountStore(src, offset, *this)) {
     LOG_GENERAL(WARNING, "Messenger::GetAccountStore failed.");
@@ -148,9 +144,7 @@ bool AccountStore::Deserialize(const bytes& src, unsigned int offset) {
 bool AccountStore::SerializeDelta() {
   LOG_MARKER();
 
-  unique_lock<mutex> g(m_mutexDelta, defer_lock);
-  shared_lock<shared_timed_mutex> g2(m_mutexPrimary, defer_lock);
-  lock(g, g2);
+  unique_lock<shared_timed_mutex> g(m_mutexDelta);
 
   m_stateDeltaSerialized.clear();
 
@@ -164,77 +158,84 @@ bool AccountStore::SerializeDelta() {
 }
 
 void AccountStore::GetSerializedDelta(bytes& dst) {
-  lock_guard<mutex> g(m_mutexDelta);
-
   dst.clear();
-
-  copy(m_stateDeltaSerialized.begin(), m_stateDeltaSerialized.end(),
-       back_inserter(dst));
+  {
+    shared_lock<shared_timed_mutex> g(m_mutexDelta);
+    copy(m_stateDeltaSerialized.begin(), m_stateDeltaSerialized.end(),
+         back_inserter(dst));
+  }
 }
 
 bool AccountStore::DeserializeDelta(const bytes& src, unsigned int offset,
                                     bool revertible) {
   LOG_MARKER();
 
-  if (revertible) {
-    unique_lock<shared_timed_mutex> g(m_mutexPrimary, defer_lock);
-    unique_lock<mutex> g2(m_mutexRevertibles, defer_lock);
-    lock(g, g2);
-
-    if (!Messenger::GetAccountStoreDelta(src, offset, *this, revertible,
-                                         false)) {
-      LOG_GENERAL(WARNING, "Messenger::GetAccountStoreDelta failed.");
-      return false;
-    }
-  } else {
-    unique_lock<shared_timed_mutex> g(m_mutexPrimary);
-
-    if (!Messenger::GetAccountStoreDelta(src, offset, *this, revertible,
-                                         false)) {
-      LOG_GENERAL(WARNING, "Messenger::GetAccountStoreDelta failed.");
-      return false;
-    }
+  if (!Messenger::GetAccountStoreDelta(src, offset, *this, revertible, false)) {
+    LOG_GENERAL(WARNING, "Messenger::GetAccountStoreDelta failed.");
+    return false;
   }
 
   return true;
 }
 
 bool AccountStore::DeserializeDeltaTemp(const bytes& src, unsigned int offset) {
-  lock_guard<mutex> g(m_mutexDelta);
   return m_accountStoreTemp->DeserializeDelta(src, offset);
 }
 
-bool AccountStore::MoveRootToDisk(const h256& root) {
-  // convert h256 to bytes
-  if (!BlockStorage::GetBlockStorage().PutStateRoot(root.asBytes())) {
-    LOG_GENERAL(INFO, "FAIL: Put state root failed " << root.hex());
+bool AccountStore::UpdateStateTrieFromTempStateDB() {
+  LOG_MARKER();
+
+  leveldb::Iterator* iter = nullptr;
+
+  while (iter == nullptr || iter->Valid()) {
+    vector<StateSharedPtr> states;
+    if (!BlockStorage::GetBlockStorage().GetTempStateInBatch(iter, states)) {
+      LOG_GENERAL(WARNING, "GetTempStateInBatch failed");
+      delete iter;
+      return false;
+    }
+    for (const auto& state : states) {
+      UpdateStateTrie(state->first, state->second);
+    }
+  }
+
+  delete iter;
+
+  if (!BlockStorage::GetBlockStorage().ResetDB(BlockStorage::TEMP_STATE)) {
+    LOG_GENERAL(WARNING, "BlockStorage::ResetDB (TEMP_STATE) failed");
     return false;
   }
   return true;
 }
 
-bool AccountStore::MoveUpdatesToDisk() {
+bool AccountStore::RetrieveFromDisk() {
   LOG_MARKER();
 
-  unique_lock<shared_timed_mutex> g(m_mutexPrimary, defer_lock);
-  unique_lock<mutex> g2(m_mutexDB, defer_lock);
-  lock(g, g2);
+  InitSoft();
+
+  return AccountStoreTrie<
+      dev::OverlayDB, std::unordered_map<Address, std::shared_ptr<Account>>>::
+      RetrieveFromDisk();
+}
+
+bool AccountStore::MoveUpdatesToDisk() {
+  LOG_MARKER();
 
   unordered_map<string, string> code_batch;
   unordered_map<string, string> initdata_batch;
 
-  for (const auto& i : *m_addressToAccount) {
-    if (i.second.isContract()) {
+  for (const auto& i : GetAccounts()) {
+    if (i.second->isContract()) {
       if (ContractStorage2::GetContractStorage()
               .GetContractCode(i.first)
               .empty()) {
         code_batch.insert({i.first.hex(), DataConversion::CharArrayToString(
-                                              i.second.GetCode())});
+                                              i.second->GetCode())});
       }
 
       if (ContractStorage2::GetContractStorage().GetInitData(i.first).empty()) {
         initdata_batch.insert({i.first.hex(), DataConversion::CharArrayToString(
-                                                  i.second.GetInitData())});
+                                                  i.second->GetInitData())});
       }
     }
   }
@@ -271,114 +272,14 @@ bool AccountStore::MoveUpdatesToDisk() {
     }
   }
 
-  try {
-    lock_guard<mutex> g(m_mutexTrie);
-    if (!m_state.db()->commit()) {
-      LOG_GENERAL(WARNING, "LevelDB commit failed");
-    }
-    if (!MoveRootToDisk(m_state.root())) {
-      LOG_GENERAL(WARNING, "MoveRootToDisk failed " << m_state.root().hex());
-      return false;
-    }
-    m_prevRoot = m_state.root();
-  } catch (const boost::exception& e) {
-    LOG_GENERAL(WARNING, "Error with AccountStore::MoveUpdatesToDisk. "
-                             << boost::diagnostic_information(e));
-    return false;
-  }
-
-  m_addressToAccount->clear();
-
-  return true;
+  return AccountStoreTrie<
+      dev::OverlayDB, std::unordered_map<Address, std::shared_ptr<Account>>>::
+      MoveUpdatesToDisk();
 }
 
-bool AccountStore::UpdateStateTrieFromTempStateDB() {
-  LOG_MARKER();
-
-  leveldb::Iterator* iter = nullptr;
-
-  while (iter == nullptr || iter->Valid()) {
-    vector<StateSharedPtr> states;
-    if (!BlockStorage::GetBlockStorage().GetTempStateInBatch(iter, states)) {
-      LOG_GENERAL(WARNING, "GetTempStateInBatch failed");
-      delete iter;
-      return false;
-    }
-    for (const auto& state : states) {
-      UpdateStateTrie(state->first, state->second);
-    }
-  }
-
-  delete iter;
-
-  if (!BlockStorage::GetBlockStorage().ResetDB(BlockStorage::TEMP_STATE)) {
-    LOG_GENERAL(WARNING, "BlockStorage::ResetDB (TEMP_STATE) failed");
-    return false;
-  }
-  return true;
-}
-
-void AccountStore::DiscardUnsavedUpdates() {
-  LOG_MARKER();
-
-  unique_lock<shared_timed_mutex> g(m_mutexPrimary, defer_lock);
-  unique_lock<mutex> g2(m_mutexDB, defer_lock);
-  lock(g, g2);
-
-  try {
-    {
-      lock_guard<mutex> g(m_mutexTrie);
-      m_state.db()->rollback();
-      m_state.setRoot(m_prevRoot);
-    }
-    m_addressToAccount->clear();
-  } catch (const boost::exception& e) {
-    LOG_GENERAL(WARNING, "Error with AccountStore::DiscardUnsavedUpdates. "
-                             << boost::diagnostic_information(e));
-  }
-}
-
-bool AccountStore::RetrieveFromDisk() {
-  LOG_MARKER();
-
-  InitSoft();
-
-  unique_lock<shared_timed_mutex> g(m_mutexPrimary, defer_lock);
-  unique_lock<mutex> g2(m_mutexDB, defer_lock);
-  lock(g, g2);
-
-  bytes rootBytes;
-  if (!BlockStorage::GetBlockStorage().GetStateRoot(rootBytes)) {
-    // To support backward compatibilty - lookup with new binary trying to
-    // recover from old database
-    if (BlockStorage::GetBlockStorage().GetMetadata(STATEROOT, rootBytes)) {
-      if (!BlockStorage::GetBlockStorage().PutStateRoot(rootBytes)) {
-        LOG_GENERAL(WARNING,
-                    "BlockStorage::PutStateRoot failed "
-                        << DataConversion::CharArrayToString(rootBytes));
-        return false;
-      }
-    } else {
-      LOG_GENERAL(WARNING, "Failed to retrieve StateRoot from disk");
-      return false;
-    }
-  }
-
-  try {
-    h256 root(rootBytes);
-    LOG_GENERAL(INFO, "StateRootHash:" << root.hex());
-    lock_guard<mutex> g(m_mutexTrie);
-    m_state.setRoot(root);
-  } catch (const boost::exception& e) {
-    LOG_GENERAL(WARNING, "Error with AccountStore::RetrieveFromDisk. "
-                             << boost::diagnostic_information(e));
-    return false;
-  }
-  return true;
-}
-
-Account* AccountStore::GetAccountTemp(const Address& address) {
-  return m_accountStoreTemp->GetAccount(address);
+std::unique_lock<std::mutex> AccountStore::GetAccountWMutexTemp(
+    const Address& address, std::shared_ptr<Account>& acc) {
+  return m_accountStoreTemp->GetAccountWMutex(address, acc);
 }
 
 bool AccountStore::UpdateAccountsTemp(const uint64_t& blockNum,
@@ -387,12 +288,6 @@ bool AccountStore::UpdateAccountsTemp(const uint64_t& blockNum,
                                       const Transaction& transaction,
                                       TransactionReceipt& receipt,
                                       TxnStatus& error_code) {
-  // LOG_MARKER();
-
-  unique_lock<shared_timed_mutex> g(m_mutexPrimary, defer_lock);
-  unique_lock<mutex> g2(m_mutexDelta, defer_lock);
-  lock(g, g2);
-
   return m_accountStoreTemp->UpdateAccounts(blockNum, numShards, isDS,
                                             transaction, receipt, error_code);
 }
@@ -400,22 +295,21 @@ bool AccountStore::UpdateAccountsTemp(const uint64_t& blockNum,
 bool AccountStore::UpdateCoinbaseTemp(const Address& rewardee,
                                       const Address& genesisAddress,
                                       const uint128_t& amount) {
-  // LOG_MARKER();
-
-  lock_guard<mutex> g(m_mutexDelta);
-
-  if (m_accountStoreTemp->GetAccount(rewardee) == nullptr) {
-    m_accountStoreTemp->AddAccount(rewardee, {0, 0});
+  {
+    std::shared_ptr<Account> acc;
+    unique_lock<std::mutex> g(GetAccountWMutexTemp(rewardee, acc));
+    if (acc == nullptr) {
+      g.unlock();
+      m_accountStoreTemp->AddAccount(rewardee, make_shared<Account>(0, 0));
+    }
   }
+
   return m_accountStoreTemp->TransferBalance(genesisAddress, rewardee, amount);
-  // Should the nonce increase ??
 }
 
 uint128_t AccountStore::GetNonceTemp(const Address& address) {
-  lock_guard<mutex> g(m_mutexDelta);
-
-  if (m_accountStoreTemp->GetAddressToAccount()->find(address) !=
-      m_accountStoreTemp->GetAddressToAccount()->end()) {
+  if (m_accountStoreTemp->GetAccounts().find(address) !=
+      m_accountStoreTemp->GetAccounts().end()) {
     return m_accountStoreTemp->GetNonce(address);
   } else {
     return this->GetNonce(address);
@@ -423,28 +317,29 @@ uint128_t AccountStore::GetNonceTemp(const Address& address) {
 }
 
 StateHash AccountStore::GetStateDeltaHash() {
-  lock_guard<mutex> g(m_mutexDelta);
-
   bool isEmpty = true;
-
-  for (unsigned char i : m_stateDeltaSerialized) {
-    if (i != 0) {
-      isEmpty = false;
-      break;
-    }
-  }
-
-  if (isEmpty) {
-    return StateHash();
-  }
-
   SHA2<HashType::HASH_VARIANT_256> sha2;
-  sha2.Update(m_stateDeltaSerialized);
+  {
+    shared_lock<shared_timed_mutex> g(m_mutexDelta);
+    for (unsigned char i : m_stateDeltaSerialized) {
+      if (i != 0) {
+        isEmpty = false;
+        break;
+      }
+    }
+
+    if (isEmpty) {
+      return StateHash();
+    }
+
+    sha2.Update(m_stateDeltaSerialized);
+  }
   return StateHash(sha2.Finalize());
 }
 
 void AccountStore::CommitTemp() {
   LOG_MARKER();
+  shared_lock<shared_timed_mutex> g(m_mutexDelta);
   if (!DeserializeDelta(m_stateDeltaSerialized, 0)) {
     LOG_GENERAL(WARNING, "DeserializeDelta failed.");
   }
@@ -455,6 +350,7 @@ void AccountStore::CommitTempRevertible() {
 
   InitRevertibles();
 
+  shared_lock<shared_timed_mutex> g(m_mutexDelta);
   if (!DeserializeDelta(m_stateDeltaSerialized, 0, true)) {
     LOG_GENERAL(WARNING, "DeserializeDelta failed.");
   }
@@ -463,16 +359,18 @@ void AccountStore::CommitTempRevertible() {
 void AccountStore::RevertCommitTemp() {
   LOG_MARKER();
 
-  unique_lock<shared_timed_mutex> g(m_mutexPrimary);
+  {
+    lock_guard<mutex> g(m_mutexRevertibles);
 
-  // Revert changed
-  for (auto const& entry : m_addressToAccountRevChanged) {
-    (*m_addressToAccount)[entry.first] = entry.second;
-    UpdateStateTrie(entry.first, entry.second);
-  }
-  for (auto const& entry : m_addressToAccountRevCreated) {
-    RemoveAccount(entry.first);
-    RemoveFromTrie(entry.first);
+    // Revert changed
+    for (auto const& entry : m_addressToAccountRevChanged) {
+      AddAccount(entry.first, entry.second, true);
+      UpdateStateTrie(entry.first, entry.second);
+    }
+    for (auto const& entry : m_addressToAccountRevCreated) {
+      RemoveAccount(entry.first);
+      RemoveFromTrie(entry.first);
+    }
   }
 
   ContractStorage2::GetContractStorage().RevertContractStates();
@@ -497,14 +395,14 @@ bool AccountStore::MigrateContractStates2(
 
     LOG_GENERAL(INFO, "Address: " << address.hex());
 
-    Account account;
-    if (!account.DeserializeBase(bytes(i.second.begin(), i.second.end()), 0)) {
+    std::shared_ptr<Account> account = std::make_shared<Account>();
+    if (!account->DeserializeBase(bytes(i.second.begin(), i.second.end()), 0)) {
       LOG_GENERAL(WARNING, "Account::DeserializeBase failed");
       return false;
     }
 
-    if (account.isContract()) {
-      account.SetAddress(address);
+    if (account->isContract()) {
+      account->SetAddress(address);
       if (!contract_address_output_dir.empty()) {
         os_1 << address.hex() << endl;
       }
@@ -521,7 +419,7 @@ bool AccountStore::MigrateContractStates2(
     bool is_library;
     uint32_t scilla_version;
     std::vector<Address> extlibs;
-    if (!account.GetContractAuxiliaries(is_library, scilla_version, extlibs)) {
+    if (!account->GetContractAuxiliaries(is_library, scilla_version, extlibs)) {
       LOG_GENERAL(WARNING, "GetScillaVersion failed");
       return false;
     }
@@ -537,8 +435,7 @@ bool AccountStore::MigrateContractStates2(
       return false;
     }
 
-    if (!ExportCreateContractFiles(account, is_library, scilla_version,
-                                   extlibs_exports)) {
+    if (!ExportCreateContractFiles(*account, is_library, extlibs_exports)) {
       LOG_GENERAL(WARNING, "ExportCreateContractFiles failed");
       return false;
     }
@@ -548,9 +445,11 @@ bool AccountStore::MigrateContractStates2(
     bool ret_checker = true;
     TransactionReceipt receipt;
     uint64_t gasRem = UINT64_MAX;
-    InvokeInterpreter(CHECKER, checkerPrint, scilla_version, is_library, gasRem,
-                      std::numeric_limits<uint128_t>::max(), ret_checker,
-                      receipt);
+    if (!InvokeInterpreter(CHECKER, checkerPrint, scilla_version, is_library,
+                           gasRem, std::numeric_limits<uint128_t>::max(),
+                           receipt)) {
+      ret_checker = false;
+    }
 
     if (!ret_checker) {
       LOG_GENERAL(WARNING, "InvokeScillaChecker failed");
@@ -579,14 +478,14 @@ bool AccountStore::MigrateContractStates2(
         Contract::ContractStorage2::GetContractStorage().GenerateStorageKey(
             address, FIELDS_MAP_DEPTH_INDICATOR, {}));
 
-    account.SetStorageRoot(dev::h256());
+    account->SetStorageRoot(dev::h256());
 
-    account.UpdateStates(address, t_metadata, toDeletes, false, false, true);
+    account->UpdateStates(address, t_metadata, toDeletes, false, false, true);
 
     this->AddAccount(address, account, true);
 
-    if (account.isContract())
-      LOG_GENERAL(INFO, "storageRoot: " << account.GetStorageRoot().hex());
+    if (account->isContract())
+      LOG_GENERAL(INFO, "storageRoot: " << account->GetStorageRoot().hex());
   }
 
   if (!contract_address_output_dir.empty()) {
